Changed the adding of the fuel data to be conditional.
That is, if the "fuel" directory exists locally, it will
be copied into the image; this is useful for the maintainers
to build the complete image.  If the "fuel" directory does
not exist, then it is skipped; this is useful for someone
just wanting to build the dockerfile locally.

Since the fuel data is really just a cache, everything will
work in both cases, it will just be faster in the former
case.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>